refactor: make type local
